At this point,
Bluetooth reveiver is pretty much done. Copy pasta the code that I need and add the relay.
If I can get the remote to send the right stuff to the reveiver, that part is almost done.
Will need to add interrupts or something when that happens. Cross that bridge when we get there.