Finally fixed it.
So the bluetooth adds some mumbo jumbo to the string I have it reading in. I needed to add a delay during the read or the string would not concat.
After figuring it out, I found that some delays were only concating half the string. In order to fix this, I increased the delay until it stopped having issues.
Since it adds the same jumbo to all incoming, I used the input as a unique string. The if statements are looking for SPECIFICALLY what the bluetooth is transmitting.
There must be some way to fix that, but for now I will adapt. The switch statement was swapped for else if's because switch statements in C/Cpp don't support strings
as inputs to switches.

Todo:
Add support to switching effects on the screen. Should be easy enough.
More comments
Add support for serial from the computer too!
Connect the actual relays instead of LED's to verify functionality of hardware.