This is the first commit in a few days I think.
I went and added a .gitignore, no idea if it works or not. I'll be looking at these things shortly.
More concerned with meeting qualifiers for the project. I have successfully figured out how to do some DSP using the STM32F4 discovery platform.
I have also successfully run the board on an external supply which is crucial to the project! Super stoked.
Time to put the hammer down some more on the dsp file.